Latest attempt to get timestamps correct with minizip-ng, but still failing
Testing has switched to a console project which took a few more steps and got most of the timestamps correct
for the big test example zip (over 1000 correct timestamps) but there were 14 files in the zip that persistently
got incorrect timestamps, just out by +1 hour. After lots of attempt to try to get raw flag fields from minizip
we'll ice this effort and try a new branch using libzip. Although not as lightweight, this is a maintained GitHub
repo that is much more likely to give us the correct timestamp behaviour for all files, as InfoWare unzip does
for this big test zip.

Note: this is not the final version of ApplyZipTimestamp.... see code in TestTimestamps22 console project for the
closest-to-working version