Migrate the unzip library to InfoZip's libzip as minizip's timestamps are unreliable
This now uses the same library as InfoZip's unzip.exe which is here: https://github.com/winlibs/libzip
In this repo now we just take the 64bit unicode builds of libzip into .lib files and directly use
the include and lib files from that build here.

This now tests ok on a number of very big zips which were unpacking with unreliable datestamps using minizip.
Cannot fault this so far.